Re-organize codebase
Source code is now in `src`, `setVersion` is now in `bin` since it's a script with side-effects and not a simple helper function.

`set-version` also accepts a version number when caleld:

```
yarn set-version 1.4.5
```

(`$TRAVIS_TAG` is still supported)
