+++ .travis.yml
@@ -3,7 +3,7 @@ language: node_js
 node_js:
 - node
 
-script: yarn && yarn test && yarn build
+script: yarn && yarn test && yarn build && yarn set-version $TRAVIS_TAG
 
 deploy:
   provider: npm
