+++ package.json
@@ -9,14 +9,14 @@
   },
   "scripts": {
     "validate-filenames": "validate-filenames",
-    "lint-ts": "tslint --project ./helpers 'helpers/**/*.ts{,x}'",
-    "check-ts": "tsc --project ./helpers --noEmit",
-    "build/ts": "tsc --project ./helpers",
-    "build/babel": "babel helpers -d helpers",
-    "build": "run-s clean build/ts build/babel set-version",
-    "set-version": "node helpers/setVersion.js",
+    "lint-ts": "tslint --project ./src 'src/**/*.ts{,x}'",
+    "check-ts": "tsc --project ./src --noEmit",
+    "build/ts": "tsc --project ./src",
+    "build/babel": "babel helpers -d helpers && babel bin -d bin",
+    "build": "run-s clean build/ts build/babel",
+    "set-version": "node bin/setVersion.js",
     "lint-staged": "lint-staged",
-    "clean": "del-cli 'helpers/**.{d.ts,js}'",
+    "clean": "del-cli helpers bin",
     "lint": "run-p -c --aggregate-output lint-ts",
     "pretest": "run-p -c --aggregate-output lint validate-filenames lint-staged",
     "test": "run-p --aggregate-output -l check-ts"
@@ -37,12 +37,13 @@
     ]
   },
   "bin": {
-    "set-version": "./helpers/setVersion.js"
+    "set-version": "./bin/setVersion.js"
   },
   "dependencies": {
     "@babel/runtime": "^7.0.0-beta.40",
     "@types/aws-lambda": "^0.0.34",
     "bluebird": "^3.5.1",
+    "common-tags": "^1.7.2",
     "glob": "^7.1.2",
     "jszip": "^3.1.4",
     "lodash": "^4.17.4",
@@ -81,6 +82,7 @@
     "@hollowverse/config": "^4.0.0",
     "@hollowverse/validate-filenames": "^1.3.9",
     "@types/bluebird": "^3.5.20",
+    "@types/common-tags": "^1.4.0",
     "@types/glob": "^5.0.32",
     "@types/jszip": "^3.1.3",
     "@types/lodash": "^4.14.74",
