+++ src/helpers/executeCommandsInParallel.ts
