+++ src/tsconfig.json
@@ -5,7 +5,7 @@
     "inlineSourceMap": true,
     "declaration": true,
     "allowJs": false,
-    "checkJs": false
-  },
-  "include": ["./**/*.ts"]
+    "checkJs": false,
+    "outDir": "../"
+  }
 }
